fowm.jl
Finished code for simulating the fowm, however the simulate_fowm function right now only deals with Choke Opening as scalar not as a function of time. Furthermore, adding a struct or a dict to the Well props and a function that calculates the pressures after solving the system of ODE is needed to do the Plots later on.